@@ -17,9 +17,9 @@
         "@babel/core": "7.16.0",
         "@craco/craco": "^7.1.0",
         "@hookform/resolvers": "^5.2.1",
-        "@testing-library/jest-dom": "^5.11.4",
-        "@testing-library/react": "^11.1.0",
-        "@testing-library/user-event": "^12.1.10",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.5.1",
         "aos": "^2.3.4",
         "axios": "^0.24.0",
         "chart.js": "^3.9.1",
@@ -27,16 +27,16 @@
         "init": "^0.1.2",
         "jspdf": "^3.0.3",
         "leaflet": "^1.9.4",
-        "react": "^17.0.2",
+        "react": "^18.3.1",
         "react-bootstrap": "^2.10.10",
         "react-chartjs-2": "^4.3.1",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.3.1",
         "react-dropzone": "^14.3.8",
         "react-helmet": "^6.1.0",
         "react-helmet-async": "^2.0.5",
         "react-hook-form": "^7.62.0",
         "react-icons": "^5.5.0",
-        "react-leaflet": "^3.2.5",
+        "react-leaflet": "^4.2.1",
         "react-router-dom": "^5.3.0",
         "react-scripts": "5.0.1",
         "react-youtube": "^10.1.0",
