@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "App";
 import "index.css";
 import reportWebVitals from "reportWebVitals";
@@ -9,11 +9,11 @@ if (typeof window !== 'undefined' && !window.process) {
   window.process = { env: {} };
 }
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root"));
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
